Git ignore the venv/ folder.
This folder is system-dependent and should be local to each user. Therefore, it does not need to be checked into git, and should be ignored.

The folder venv/ contains the "special" versions of python and pip which only install and use packages from the virtual environment. It also contains the packages themselves.