Add the first version of the yahtzee program.
This is the simplest possible proof-of-concept. It makes a request to an API which returns the roll of a dice, and prints out the result. The API does not need any parameters, and the text which is returned is printed out as-is.

As we used the 'requests' Python library to make the HTTP request, this has been added to the file 'requirements.txt'.